requirements.txt
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.2
cryptojwt==1.6.1
filelock==3.6.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
mysql-connector-python==8.0.25
oidcmsg==1.6.0
oidcrp==2.1.4
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
protobuf==3.19.4
py==1.11.0
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.18.0
PyYAML==6.0
readerwriterlock==1.0.9
requests==2.27.1
responses==0.20.0
six==1.16.0
SQLAlchemy==1.4.20
toml==0.10.2
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.0.3
